+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- silo-core: count for fractions when calculate maxBorrow
 
 ## [3.3.0] - 2025-05-01
 - silo-vaults: SiloIncentivesControllerCLDeployer helper contract to deploy claiming logics